cloning div
was able to make the item button create a clone of the table and place it directly below. Having trouble trying to figure out how to change the ID of the clone dynamically. I want the new ID to be named something like 'userData1', 'userData2', and so on. I will come back to this part of the webapp