clone div and create ne id dynamically
I figured out how to create a dynamic ID everytime the user clickes the 'new entry' button.

I created a global variable 'cloneCounter' and set its' value to 0. In the New entry function, everytime the button is pressed a clone of the 'userDataRow' div is created and then the cloneCounter gets plus 1. After the cloneCounter gets a its' new value the id is set to 'userData' + cloneCounter.